feat(PPDSC-2064): feat-ppdse/2064 ref on menuitem
